% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_measure.R
\name{average_measure}
\alias{average_measure}
\title{Create a summary COVID19 data set for the entire USA}
\usage{
average_measure(data, measure)
}
\arguments{
\item{data}{A data set to be summarised}

\item{measure}{The variable to summarise the data on, can be as a character vector. Can be \code{"tot_cases"} or \verb{"tot_death}.}
}
\value{
A summarised data set comprising 251 observations (one for each day) and 3 variables dependent on what \code{measure} is chosen.
}
\description{
Create new data set comprising average daily counts for each different \code{measure} of COVID19 in the USA.
}
\examples{
average_measure(usa_covid_data, "tot_cases")
average_measure(usa_covid_data, "tot_death")

}
