% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input.R
\name{create_input}
\alias{create_input}
\title{Create inputs for shiny application}
\usage{
create_input(input)
}
\arguments{
\item{input}{The variable chosen for the input as a character vector. Can be \code{"measure"}, \code{"state"}, or \code{"date"}.}
}
\value{
A shiny control widget used to specify inputs for the visualisations in the application
}
\description{
Create chosen shiny inputs for the different variables contained in the data set \code{usa_covid_data} and \code{usa_state_map}.
}
\examples{
create_input("measure")
create_input("state")
create_input("date")

}
