% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_covid_data.R
\docType{data}
\name{usa_covid_data}
\alias{usa_covid_data}
\title{Daily COVID19 cases and deaths in the USA by state}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15060 rows and 13 columns.
}
\source{
Centers for Disease Control and Prevention
}
\usage{
usa_covid_data
}
\description{
This dataset contains COVID19 related data from 60 states and territories of the USA from the 22nd of January to the 28th of September 2020. In total there are 15,060 rows across 13 variables.
}
\keyword{datasets}
