% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_state_map.R
\docType{data}
\name{usa_state_map}
\alias{usa_state_map}
\title{State borders for the continental United States}
\format{
An object of class \code{data.frame} with 15537 rows and 6 columns.
}
\source{
US Department of the Census
}
\usage{
usa_state_map
}
\description{
This dataset contains spatial data provided in a tabular format, producing a map of the United States mainland and is loaded directly from the maps package.
}
\keyword{datasets}
