% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_table.R
\name{present_table}
\alias{present_table}
\title{Present a stylish table}
\usage{
present_table(data, measure)
}
\arguments{
\item{data}{The data table to be formatted}

\item{measure}{The measure chosen for the heading of the table as a character vector, can be \code{"cases"} or \code{"deaths"}}
}
\value{
A formatted kable of the original data set entered
}
\description{
Creates a formatted and styled table to display COVID19 data using the kableExtra package
}
\examples{
present_table(combined, "cases")
present_table(combined, "deaths")

}
