% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_states.R
\name{random_states}
\alias{random_states}
\title{Compare two states}
\usage{
random_states()
}
\value{
A collection of two state names returned from the \code{usa_state_map} data set.
}
\description{
This function is used in the shiny application to provide two random state names each time it is loaded for the default values.
}
\examples{
random_states()

}
